Refrector CSS styles for card container and image
This commit lowers the max-width of the card container from
63.75rem to 30rem for mobile screen sizes and applies max-width
of 63.75 for desktop version.

A margin of 1.5rem is applied to the main element.

mix-blend-mode is applied to the image and opacity of 0.75
also applied.

This commit improves responsiveness between mobile and desktop
versions of the component, applies a more appropriate max-width
for mobile viewing and adds a decorative color effect to the image.